backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
alembic==1.13.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
redis==5.0.8
httpx==0.27.2
anchorpy==0.21.0
solana==0.36.1
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
qrcode[pil]==7.4.2
python-multipart==0.0.6

